audiolat: Add midi playout functionality
Adds "-e midi 1" to command line to enable midi.
A list of available devices will be shown on screen.
Test starts when one device is selected by clicking on it.
Test will continue until the timeout has passed and
a recorded file can be pulled the normal way.

The find_transient.py script can be used to find transients
and signal in a wave file. The following call is used to
find a transient trigger before the marker signal.
>find_transients.py -m end.wav capture.wav

-l means marker leading and is used to fing a transient
that is after the marker signal.
>find_transients.py -l -m begin.wav capture.wav

Signed-off-by: Johan Blome <jblome@fb.com>