audiolat: Fixes for Pixel3 and AAudio
AAudio path was choosing the wrong path in later code for AAudio
causing it to fail when openign a record stream.
Different "usage" settings gives very different output and a
additional parameter have been added to the "find_transients.py" script.
--limit_marker SEC
Limits the length of the marker and the matching to SEC.